fix: docker setup
