+++ btc-deploy/Ordhook.toml
@@ -1,0 +1,39 @@
+[storage]
+working_dir = "ordhook"
+
+# The Http Api allows you to register / deregister
+# dynamically predicates.
+# Disable by default.
+#
+[http_api]
+http_port = 20456
+
+[network]
+mode = "devnet"
+bitcoind_rpc_url = "http://bitcoind:18443"
+bitcoind_rpc_username = "ic-btc-integration"
+bitcoind_rpc_password = "QPQiNaph19FqUsCrBRN0FII7lyM26B51fAMeBQzCb-E="
+# Bitcoin block events can be received by Chainhook
+# either through a Bitcoin node's ZeroMQ interface,
+# or through the Stacks node. Zmq is being
+# used by default:
+bitcoind_zmq_url = "tcp://bitcoind:28332"
+# but stacks can also be used:
+# stacks_node_rpc_url = "http://0.0.0.0:20443"
+
+[resources]
+ulimit = 2048
+cpu_core_available = 16
+memory_available = 32
+bitcoind_rpc_threads = 4
+bitcoind_rpc_timeout = 15
+expected_observers_count = 1
+
+# Disable the following section if the state
+# must be built locally
+[snapshot]
+download_url = "https://archive.hiro.so/mainnet/ordhook/mainnet-ordhook-sqlite-latest"
+
+[logs]
+ordinals_internals = true
+chainhook_internals = true
