+++ btc-deploy/docker-compose.yml
@@ -26,10 +26,45 @@ services:
       bitcoind:
         condition: service_healthy
       postgres:
-        condition: service_healthy
+        condition: service_started
     ports:
       - "8004:3000"
       - "8005:8005"
+
+  ordhook:
+    container_name: ordhook
+    build:
+      context: https://github.com/hirosystems/ordhook.git
+      dockerfile: dockerfiles/components/ordhook.dockerfile
+    ports:
+      - "20456:20456"
+    command:
+      [
+        "service",
+        "start",
+        "--post-to=http://hiro-ordinals-api:3099/payload",
+        "--auth-token=1",
+        "--config-path=/Ordhook.toml",
+      ]
+    volumes:
+      - ./Ordhook.toml:/Ordhook.toml
+    depends_on:
+      hiro-ordinals-api:
+        condition: service_started
+
+  ordinals-explorer:
+    build:
+      context: .
+      dockerfile: ordinals-explorer.dockerfile
+    env_file:
+      - ./ordinals-explorer.env
+    ports:
+      - "8006:3001"
+    volumes:
+      - ./ordinals-explorer.env:/usr/src/ordinals-explorer/.env.production
+    depends_on:
+      hiro-ordinals-api:
+        condition: service_started
 
   ord:
     container_name: ord
@@ -44,9 +79,12 @@ services:
       - "8001:8001"
 
   postgres:
+    container_name: postgres
     image: "postgres:15"
     ports:
       - "5432:5432"
+    volumes:
+      - db-data:/var/lib/postgresql/data
     environment:
       POSTGRES_USER: postgres
       POSTGRES_PASSWORD: postgres
@@ -62,3 +100,6 @@ services:
       - "8002:8002"
     extra_hosts:
       - "host.docker.internal:host-gateway"
+
+volumes:
+  db-data:
