+++ btc-deploy/ordinals-explorer.env
@@ -1,0 +1,3 @@
+PORT=3001
+NEXT_PUBLIC_API_URL=http://0.0.0.0:8004/ordinals/v1
+NEXT_PUBLIC_API_BETA_URL=http://0.0.0.0:8004/ordinals/v1
