+++ src/integration-tests/tests/utils/mod.rs
@@ -12,6 +12,8 @@ pub mod btc;
 pub mod btc_rpc_client;
 pub mod error;
 #[cfg(feature = "dfx_tests")]
+pub mod hiro_ordinals_client;
+#[cfg(feature = "dfx_tests")]
 pub mod ord_client;
 #[cfg(feature = "dfx_tests")]
 pub mod rune_helper;
